Fixed compilation under gcc. Fixed warnings under gcc. Overloaded operator = is called when it's used in variable definition. Fixed member function call of variable that is a reference to a type with member functions. Added error if one of ternary operator ?: result was void. Variable info tree is more error-prone (but not entirely :( ). Added unit tests for overloaded set and modify operators, member function call (for the case above), double in-place array with integer elements, function prototypes, member function call from member function and failure tests. Fixed some failure tests that failed for wrong reasons.

+++ NULLC/StdLib.cpp
@@ -18,7 +18,7 @@ namespace NULLC
 
 void* NULLC::AllocObject(int size)
 {
-	if(size >= minGlobalBlockSize)
+	if((unsigned int)size >= minGlobalBlockSize)
 	{
 		globalObjects.push_back(new char[size]);
 		memset(globalObjects.back(), 0, size);
